move stuff to shared
