+++ lambdas/s3hash/src/t4_lambda_s3hash/__init__.py
@@ -65,20 +65,6 @@ async def aio_context(credentials: AWSCredentials):
 
 class Checksum(ChecksumBase):
     @classmethod
-    def sha256(cls, value: bytes):
-        return cls(value=value.hex(), type=ChecksumType.SHA256)
-
-    @classmethod
-    def sha256_chunked(cls, value: bytes):
-        return cls(value=base64.b64encode(value).decode(), type=ChecksumType.SHA256_CHUNKED)
-
-    @classmethod
-    def for_parts(cls, checksums: T.Sequence[bytes]):
-        return cls.sha256_chunked(hash_parts(checksums))
-
-    _EMPTY_HASH = hashlib.sha256().digest()
-
-    @classmethod
     def empty(cls):
         return cls.sha256_chunked(cls._EMPTY_HASH) if CHUNKED_CHECKSUMS else cls.sha256(cls._EMPTY_HASH)
 
@@ -182,10 +168,6 @@ def get_compliant_checksum(attrs: GetObjectAttributesOutputTypeDef) -> T.Optiona
         return Checksum.sha256_chunked(base64.b64decode(checksum_value))
 
     return None
-
-
-def hash_parts(parts: T.Sequence[bytes]) -> bytes:
-    return hashlib.sha256(b"".join(parts)).digest()
 
 
 class PartDef(pydantic.v1.BaseModel):
